Migrate sqs metricset to use cloudwatch input as light weight module (#26117)
(cherry picked from commit adbb1a3d8cb524ff1819debd23604d71a4cc9131)
