+++ x-pack/metricbeat/include/list.go
@@ -15,7 +15,6 @@ import (
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/aws/billing"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/aws/cloudwatch"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/aws/rds"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/aws/sqs"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/awsfargate"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/awsfargate/task_stats"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/azure"
