+++ x-pack/metricbeat/module/aws/sqs/_meta/data.json
@@ -1,6 +1,12 @@
 {
     "@timestamp": "2017-10-12T08:05:34.853Z",
     "aws": {
+        "cloudwatch": {
+            "namespace": "AWS/SQS"
+        },
+        "dimensions": {
+            "QueueName": "filebeat-aws-elb-test"
+        },
         "sqs": {
             "empty_receives": 0,
             "messages": {
@@ -8,25 +14,28 @@
                 "deleted": 0,
                 "not_visible": 0,
                 "received": 0,
-                "sent": 0,
-                "visible": 2
+                "sent": 0.2857142857142857,
+                "visible": 827.8
             },
             "oldest_message_age": {
-                "sec": 78494
+                "sec": 345606.4
             },
-            "queue": {
-                "name": "test-s3-notification"
-            },
-            "sent_message_size": {}
+            "queuename": "filebeat-aws-elb-test",
+            "sent_message_size": {
+                "bytes": 1006.5
+            }
+        },
+        "tags": {
+            "created-by": "kaiyan"
         }
     },
     "cloud": {
         "account": {
-            "id": "627959692251",
-            "name": "elastic-test"
+            "id": "428152502467",
+            "name": "elastic-beats"
         },
         "provider": "aws",
-        "region": "ap-southeast-1"
+        "region": "eu-central-1"
     },
     "event": {
         "dataset": "aws.sqs",
